Unconstrain other parameters
