+++ bin/unconstrained_sandbox.jl
@@ -46,7 +46,6 @@ end
 
 #blob, mp_init, body = SampleData.gen_sample_star_dataset();
 blob, mp_init, body = SampleData.gen_sample_galaxy_dataset();
-#blob, mp_init, body = SampleData.gen_three_body_dataset();
 
 mp_original = deepcopy(mp_init)
 mp_free = deepcopy(mp_init)
@@ -77,15 +76,15 @@ compare_solutions(mp_rect, mp_free, res_rect_max_f, res_free_max_f, res_rect_ite
 
 blob, mp_init, body = SampleData.gen_three_body_dataset();
 
-mp_original = deepcopy(mp_init)
-mp_free = deepcopy(mp_init)
-mp_rect = deepcopy(mp_init)
+mp_original = deepcopy(mp_init);
+mp_free = deepcopy(mp_init);
+mp_rect = deepcopy(mp_init);
 
 # Optimize
-omitted_ids = [ids_free.kappa[:], ids_free.lambda[:], ids_free.zeta[:] ]
+omitted_ids = [ids_free.kappa[:], ids_free.lambda[:], ids_free.zeta[:] ];
 
-lbs_free, ubs_free = OptimizeElbo.get_nlopt_unconstrained_bounds(mp_original.vp, omitted_ids, free_transform)
-lbs_rect, ubs_rect= OptimizeElbo.get_nlopt_unconstrained_bounds(mp_original.vp, omitted_ids, rect_transform)
+lbs_free, ubs_free = OptimizeElbo.get_nlopt_unconstrained_bounds(mp_original.vp, omitted_ids, free_transform);
+lbs_rect, ubs_rect= OptimizeElbo.get_nlopt_unconstrained_bounds(mp_original.vp, omitted_ids, rect_transform);
 
 mp_rect_three = deepcopy(mp_init)
 res_rect_iter_count_three, res_rect_max_f_three, res_rect_max_x, res_rect_ret =
